Extract glob tests to a separate collection
